* AppController.m ([AppController -compareFileToCVS:sender]): Get the path to the cvs executable. * DiffView.m ([DiffView -tile]): ceilf->ceil * DiffWindowController.m (-initWithFilename:andTempFilename:): Switch filenames so CVS version is on the left.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/easydiff/trunk@21308 72102866-910b-0410-8b05-ffd578937521

+++ DiffView.m
@@ -196,7 +196,7 @@
 
   width = (size.width - 80) / 2;
 
-  width = ceilf(width);
+  width = ceil(width);
   
   [leftView setFrame: 
 	      NSMakeRect(0, 0,
